requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
certifi==2025.11.12
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.124.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.7.4
pydantic-settings==2.12.0
pydantic_core==2.18.4
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
SQLAlchemy==2.0.45
starlette==0.50.0
Telethon==1.42.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
